// src/models/sign_up.ts
export class SignupDetails { 
    DomainID:string;
    CategoryId:string;
    Category:string; 
    BrandName:string; 
    About:string; 
    Country:string; 
    EmailId: string;    
    PhoneNumber: number;
    CountryCode: string;
    Street: string;
    City: string;
    State: string;
    PostalCode: string;
    UserName: string;
    Designation: string;
    UserEmailId: string;
    RegBusinessName: string;

    RegisteredType: string; 
    AccountPassword: string; 

  }

  export class BankDetails {
     
  }


  